XCW/Hoshino/hoshino/modules/pcrjjc2_xcw/requirement.txt
msgpack>=1.0.2
pycryptodome>=3.9.9